Move feature stubbing to test helpers
